Data preprocessing

Documentaion for ProRF's data preprocessing function.

ProRF.data_preprocess_index — Function

data_preprocess_index(in_fasta_loc::String;
                      target_rate::Float64=0.3,
                      val_mode::Bool=false)

Examples

julia> Find, Lind = data_preprocess_index("Data/algpdata.fasta", val_mode=true);

julia> @printf "%d %d\n" Find Lind
26 492

Analyze aligned sequence, and get index to slice aligned sequence.

Return front, last index whose gap ratio is smaller than target_rate at each ends, then display partial of both ends when val_mode is off.

Arguments

in_fasta_loc::String : location of .fasta file

target_rate::Float64 : baseline for gap ratio

val_mode::Bool : when val_mode is true, function don't display anything.

ProRF.data_preprocess_fill — Function

data_preprocess_fill(front_ind::Int, last_ind::Int,
                     in_fasta_loc::String,
                     newick_loc::String,
                     out_fasta_loc::String;
                     val_mode::Bool=false)

Examples

julia> data_preprocess_fill(Find, Lind,
                            "Data/algpdata.fasta",
                            "Data/Mega/ealtreedata.nwk",
                            "Data/jealgpdata.fasta",
                            val_mode=true);
<Closest Target --> Main>
XP_045911656.1 --> AAR20843.1
XP_042307646.1 --> XP_042700069.1
XP_006741075.1 --> ABD77268.1
AAL66372.1 --> AAH29754.1
XP_006741075.1 --> AAA31017.1
XP_006741075.1 --> ABD77263.1
XP_031465919.1 --> XP_046760974.1
XP_032600743.1 --> XP_014813383.1
XP_033927480.1 --> PKK17119.1

Fill aligned partial .fasta file with nearest data without having a gap by .nwk file, then display edited data.

Make sure that id of .fasta file same as id of .nwk file.

Arguments

front_ind::Int, last_ind::Int : front, last index by data_preprocess_index or user defined.

in_fasta_loc::String : location of .fasta file.

newick_loc::String : location of .fasta file.

out_fasta_loc::String : location of output .fasta file.

val_mode::Bool : when val_mode is true, function don't display anything.