Toolbox functions

Documentaion for ProRF's useful function to perform data preprocessing or random forest.

Random Forest

ProRF.blosum_matrix — Function

blosum_matrix(num::Int)

Examples

julia> blosum_matrix(90)['A', 'A']
5

julia> blosum_matrix(82)
ERROR: There are no Matrix such as BLOSUM82

Return correct BioAlignments BLOSUM matrix.

Possible matrices are BLOSUM45, BLOSUM50, BLOSUM62, BLOSUM80, BLOSUM90.

ProRF.train_test_split — Function

train_test_split(X::Matrix{Float64}, Y::Vector{Float64};
                 test_size::Float64=0.3, 
                 data_state::UInt64=@seed_u64)

Examples

julia> x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2);

Split X, Y data to train, test set.

Arguments

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

test_size::Float64 : size of test set.

data_state::UInt64 : seed used to split data.

ProRF.nrmse — Function

nrmse(pre::Vector{Float64}, tru::Vector{Float64})

Compute normalized root mean square error with predict value and true value.

nrmse(regr::RandomForestRegressor, X::Matrix{Float64}, tru::Vector{Float64})

Compute normalized root mean square error with regression model, X data and true value.

ProRF.save_model — Function

save_model(model_loc::String, regr::RandomForestRegressor)

Examples

julia> M = rf_model(X, Y, 5, 300, val_mode=true);

julia> save_model("model.jld", M);
ERROR: model.jld is not .jld2 file

julia> save_model("model.jld2", M);

Save RandomForestRegressor model using JLD2, make sure filename extension set to .jld2.

ProRF.load_model — Function

load_model(model_loc::String)

Examples

julia> M = load_model("model.jld2");

julia> X, Y, L = get_data(R, 9, 'E');

julia> @printf "Total NRMSE : %.6f\n" nrmse(M, X, Y)
Total NRMSE : 0.136494

Load RandomForestRegressor model using JLD2, make sure filename extension set to .jld2.

ProRF.parallel_predict — Function

parallel_predict(regr::RandomForestRegressor, X::Matrix{Float64})

Execute DecisionTree.predict(regr, X) in parallel.

parallel_predict(regr::RandomForestRegressor, X::Matrix{Float64})

Get raw sequence vector and L data to make X data and execute DecisionTree.predict(regr, X) in parallel.

Data preprocessing

ProRF.view_sequence — Function

view_sequence(fasta_loc::String, amino_loc::Int=1;
              fontsize::Int=9,
              seq_width::Int=800,
              save::Bool=false)

Display .fasta file sequence using PyCall.

Made it by referring to bokeh sequence aligner visualization program.

Arguments

fasta_loc::String : location of .fasta file.

amino_loc::Int : start index for .fasta file (when value is not determined, set to 1).

fontsize::Int : font size of sequence.

seq_width::Int : gap width between sequences.

save::Bool : save .html viewer.

view_sequence(R::AbstractRF;
              fontsize::Int=9,
              seq_width::Int=800,
              save::Bool=false)

Display .fasta file sequence exists at a location in the AbstractRF object using PyCall.

Made it by referring to bokeh sequence aligner visualization program.

Arguments

R::AbstractRF : for both RF and RFI.

fontsize::Int : font size of sequence.

seq_width::Int : gap width between sequences.

save::Bool : save .html viewer.

ProRF.view_mutation — Function

view_mutation(fasta_loc::String)

Analyze data from .fasta file location of AbstractRF, then draw histogram, line graph about the mutation distribution.

view_mutation(R::AbstractRF)

Analyze data from .fasta file, then draw histogram, line graph about the mutation distribution.

Others

ProRF._julia_interactive — Constant

Check julia currently running is interactive or not.

You can change the value in julia_isinteractive.

ProRF.julia_isinteractive — Function

julia_isinteractive()

Check _julia_interactive value.

julia_isinteractive(x::Bool)

Set _julia_interactive value.

ProRF.@seed_i — Macro

Return non-negative Int range integer MersenneTwister RNG object seed. keep in mind that when macro executed, the seed is initialized.

ProRF.@seed_u64 — Macro

Return UInt64 range integer MersenneTwister RNG object seed. keep in mind that when macro executed, the seed is initialized.

ProRF.@show_pyplot — Macro

When _julia_interactive is on, execute display(gcf()) or _julia_interactive is off, execute show() and wait until the user inputs enter.