Random Forest

Documentaion for ProRF's functions performing a random forest.

Main Random Forest

ProRF.AbstractRF — Type

Supertype for RF, RFI.

ProRF.AbstractRFI — Type

Supertype for RFI.

ProRF.RF — Type

RF(dataset_loc::String)

RF(fasta_loc::String, data_loc::String)

RF(fasta_loc::String, data_loc::String, amino_loc::Union{Int, Vector{Int}})

struct for Main Random Forest

Examples

julia> R = RF("Data/rgpdata.fasta", "Data/rdata.xlsx");

Fields

dataset_loc::String : location of dataset, .fasta, .xlsx file name must be data. Also, you can designate amino_loc through index.txt. See example folder.

fasta_loc::String : location of .fasta file.

data_loc::String : location of .xlsx file.

amino_loc::Union{Int, Vector{Int}} : start index or total index for amino acid (when value is not determined, set to 1).

ProRF.RFI — Type

RFI(R::AbstractRF,
    nfeat::StepRange{Int, Int}, ntree::StepRange{Int, Int})

RFI(dataset_loc::String,
    nfeat::StepRange{Int, Int}, ntree::StepRange{Int, Int}))

RFI(fasta_loc::String, data_loc::String,
    nfeat::StepRange{Int, Int}, ntree::StepRange{Int, Int})

RFI(fasta_loc::String, data_loc::String, amino_loc::Union{Int, Vector{Int}}
    nfeat::StepRange{Int, Int}, ntree::StepRange{Int, Int})

struct for Random Forest Iteration.

Examples

julia> RI = RFI("Data/rgpdata.fasta", "Data/rdata.xlsx", 2:1:10, 100:10:500);

Fields

dataset_loc::String : location of dataset, .fasta, .xlsx file name must be data. Also, you can designate amino_loc through index.txt. See example folder.

fasta_loc::String : location of .fasta file.

data_loc::String : location of .xlsx file.

amino_loc::Union{Int, Vector{Int}} : start index or total index for amino acid (when value is not determined, set to 1).

nfeat::StepRange{Int, Int} : range of the number of selected features.

ntree::StepRange{Int, Int} : range of the number of trees.

ProRF.get_data — Function

get_data(R::AbstractRF, ami_arr::Int, excel_col::Char;
         norm::Bool=false, convert::Dict{Char, Float64}=ProRF.volume,
         sheet::String="Sheet1", title::Bool=true)

get_data(R::AbstractRF, excel_col::Char;
         norm::Bool=false, convert::Dict{Char, Float64}=ProRF.volume,
         sheet::String="Sheet1", title::Bool=true)

Examples

julia> X, Y, L = get_data(R, 9, 'E');

Get data from .fasta file by converting selected dictionary and .xlsx file at certain sheet and column.

Arguments

R::AbstractRF : for both RF and RFI.

ami_arr::Int : baseline for total number of mutations in samples at one location (when value is not determined, set to 1).

excel_col::Char : column character for .xlsx file.

norm::Bool : execute min-max normalization.

convert::Dict{Char, Float64} : Convert dictionary that turns amnio acid to value

sheet::String : .xlsx data sheet name

title::Bool : when .xlsx have a header row, turn on title.

Return

X::Matrix{Float64} : independent variables data matrix.

Y::Vector{Float64} : dependent variable data vector.

L::Vector{Int} : raw sequence index vector.

ProRF.get_reg_importance — Function

get_reg_importance(R::AbstractRF, X::Matrix{Float64}, Y::Vector{Float64},
                   L::Vector{Int}, feat::Int, tree::Int;
                   val_mode::Bool=false, test_size::Float64=0.3,
                   nbin::Int=200, show_number::Int=20, imp_iter::Int=60,
                   max_depth::Int=-1,
                   min_samples_leaf::Int=1,
                   min_samples_split::Int=2,
                   data_state::UInt64=@seed,
                   learn_state::UInt64=@seed,
                   imp_state::UInt64=@seed)

Examples

julia> M, F = get_reg_importance(R, X, Y, L, 6, 800);

julia> @printf "Total NRMSE : %.6f\n" nrmse(M, X, Y)
Total NRMSE : 0.136494

Caculate regression model and feature importance, then draw random forest result and feature importance list.

Arguments

R::AbstractRF : for both RF and RFI.

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

L::Vector{Int} : L data.

feat::Int : number of selected features.

tree::Int : number of trees.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

nbin::Int : the number of bins for each two dimensions to execute kernel density estimation.

show_number::Int : number of locations to show importance.

imp_iter::Int : number of times to repeat to caculate a feature importance.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

data_state::UInt64 : seed used to split data.

learn_state::UInt64 : seed used to caculate a regression model.

imp_state::UInt64 : seed used to caculate a feature importance.

ProRF.rf_importance — Function

rf_importance(R::AbstractRF, regr::RandomForestRegressor,
              X::Matrix{Float64}, L::Vector{Int};
              val_mode::Bool=false,
              show_number::Int=20, imp_iter::Int=60,
              imp_state::UInt64=@seed)

Examples

julia> F = rf_importance(R, M, X, L, show_number=15);

Caculate feature importance for a target model, then draw feature importance list.

Arguments

R::AbstractRF : for both RF and RFI.

regr::RandomForestRegressor : target regression model.

X::Matrix{Float64} : X data.

L::Vector{Int} : L data.

val_mode::Bool : when val_mode is true, function don't display anything.

show_number::Int : number of locations to show importance.

imp_iter::Int : number of times to repeat to caculate a feature importance.

imp_state::UInt64 : seed used to caculate a feature importance.

ProRF.rf_model — Function

rf_model(X::Matrix{Float64}, Y::Vector{Float64}, feat::Int, tree::Int;
         val_mode::Bool=false, test_size::Float64=0.3, nbin::Int=200,
         max_depth::Int=-1,
         min_samples_leaf::Int=1,
         min_samples_split::Int=2,
         data_state::UInt64=@seed, 
         learn_state::UInt64=@seed)

Examples

julia> M = rf_model(X, Y, 6, 800);

Caculate regression model, then draw random forest result.

Arguments

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

feat::Int : number of selected features.

tree::Int : number of trees.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

nbin::Int : the number of bins for each two dimensions to execute kernel density estimation.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

data_state::UInt64 : seed used to split data.

learn_state::UInt64 : seed used to caculate a regression model.

ProRF.rf_nrmse — Function

rf_nrmse(X::Matrix{Float64}, Y::Vector{Float64}, feat::Int, tree::Int;
         val_mode::Bool=false, test_size::Float64=0.3, nbin::Int=200,
         max_depth::Int=-1,
         min_samples_leaf::Int=1,
         min_samples_split::Int=2,
         data_state::UInt64=@seed, 
         learn_state::UInt64=@seed)

Examples

julia> M, NE = rf_nrmse(X, Y, 6, 800);

Caculate normalized root mean square error, then draw random forest result.

Arguments

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

feat::Int : number of selected features.

tree::Int : number of trees.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

nbin::Int : the number of bins for each two dimensions to execute kernel density estimation.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

data_state::UInt64 : seed used to split data.

learn_state::UInt64 : seed used to caculate a regression model.

ProRF.iter_get_reg_importance — Function

iter_get_reg_importance(R::AbstractRF, X::Matrix{Float64}, Y::Vector{Float64},
                        L::Vector{Int},
                        feat::Int, tree::Int,
                        data_iter::Int, learn_iter::Int;
                        val_mode::Bool=false, test_size::Float64=0.3,
                        show_number::Int=20, imp_iter::Int=60,
                        max_depth::Int=-1,
                        min_samples_leaf::Int=1,
                        min_samples_split::Int=2,
                        data_state_seed::UInt64=@seed,
                        learn_state_seed::UInt64=@seed,
                        imp_state::UInt64=@seed)

Examples

julia> MF, SF = iter_get_reg_importance(R, X, Y, L, 3, 700, 10, 10);

Calculate feature importance by repeating data_iter::Int times about data seed and learn_iter::Int times about model seed, then draw feature importance list inclding standard deviation.

Returns the mean and standard deviation of feature importance.

Arguments

R::AbstractRF : for both RF and RFI.

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

L::Vector{Int} : L data.

feat::Int : number of selected features.

tree::Int : number of trees.

data_iter::Int : number of operations iterations for seed used to split data.

learn_iter::Int : number of operations iterations for seed used to caculate a regression model.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

show_number::Int : number of locations to show importance.

imp_iter::Int : number of times to repeat to caculate a feature importance.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

data_state_seed::UInt64 : seed used to generate seed used to split data.

learn_state_seed::UInt64 : seed used to generate seed used to caculate a regression model.

imp_state::UInt64 : seed used to caculate a feature importance.

ProRF.view_result — Function

view_result(regr::RandomForestRegressor, X::Matrix{Float64}, Y::Vector{Float64};
            nbin::Int=200)

Draw random forest result and return normalized root mean square error with X, Y data and regression model.

view_result(regr::RandomForestRegressor, X::Matrix{Float64}, Y::Vector{Float64},
            data_state::UInt64;
            test_size::Float64=0.3, nbin::Int=200, test_mode::Bool=true)

Draw test or train set random forest result and return normalized root mean square error with regression model, X, Y data and seed.

view_result(pre::Vector{Float64}, tru::Vector{Float64}; nbin::Int=200)

Draw random forest result and return normalized root mean square error with predict value and true value.

view_result(pre::Vector{Float64}, tru::Vector{Float64}, data_state::UInt64;
            test_size::Float64=0.3, nbin::Int=200, test_mode::Bool=true)

Draw test or train set random forest result and return normalized root mean square error with predict, true value and seed.

ProRF.view_importance — Function

view_importance(R::AbstractRF, L::Vector{Int},
                F::Vector{Float64}; show_number::Int=20)

Examples

julia> view_importance(R, L, F);

julia> view_importance(R, L, MF, show_number=30);

Draw feature importance list.

Arguments

R::AbstractRF : for both RF and RFI.

L::Vector{Int} : L data.

F::Vector{Float64} : feature importance vector.

show_number::Int : number of locations to show importance.

view_importance(R::AbstractRF, L::Vector{Int},
                MF::Vector{Float64}, SF::Vector{Float64};
                show_number::Int=20)

Examples

julia> view_importance(R, L, MF, SF, show_number=30);

Draw feature importance list with standard deviation.

Arguments

R::AbstractRF : for both RF and RFI.

L::Vector{Int} : L data.

MF::Vector{Float64} : mean feature importance vector.

SF::Vector{Float64} : standard deviation feature importance vector.

show_number::Int : number of locations to show importance.

Random Forest Iteration

ProRF.get_reg_value — Function

get_reg_value(RI::AbstractRFI, X::Matrix{Float64}, Y::Vector{Float64};
              val_mode::Bool=false, test_size::Float64=0.3,
              max_depth::Int=-1,
              min_samples_leaf::Int=1,
              min_samples_split::Int=2,
              data_state::UInt64=@seed,
              learn_state::UInt64=@seed)

Examples

julia> Z = get_reg_value(RI, X, Y, val_mode=true);

Calculate nrmse value for each nfeat, ntree condition, then draw nrmse value 3D graph.

Arguments

RI::AbstractRFI : for only RFI.

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

data_state::UInt64 : seed used to split data.

learn_state::UInt64 : seed used to caculate a regression model.

ProRF.get_reg_value_loc — Function

get_reg_value_loc(RI::AbstractRFI, Z::Matrix{Float64})

Examples

julia> @printf "%d %d\n" pf.get_reg_value_loc(RI, Z)...
7 130

Returns the best arguemnts depending on the nrmse value.

Return

Tuple{Int, Int} : best arguemnts tuple

Int : number of selected features.

Int : number of trees.

ProRF.iter_get_reg_value — Function

iter_get_reg_value(RI::AbstractRFI, X::Matrix{Float64}, Y::Vector{Float64}, iter::Int;
                   val_mode::Bool=false, test_size::Float64=0.3,
                   max_depth::Int=-1,
                   min_samples_leaf::Int=1,
                   min_samples_split::Int=2,
                   learn_state::UInt64=@seed,
                   data_state_seed::UInt64=@seed)

Examples

julia> MZ, SZ = iter_get_reg_value(RI, X, Y, 10, val_mode=true);

Calculate nrmse value for each nfeat, ntree condition by repeating iter::Int times with a fixed data seed, then draw both mean and standard deviation nrmse value 3D graph.

Returns the mean and standard deviation of nrmse value.

Arguments

RI::AbstractRFI : for only RFI.

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

val_mode::Bool : when val_mode is true, function don't display anything.

test_size::Float64 : size of test set.

max_depth::Int : maximum depth of the tree.

min_samples_leaf::Int : minimum number of samples required to be at a leaf node.

min_samples_split::Int : minimum number of samples required to split an internal node.

learn_state::UInt64 : seed used to caculate a regression model.

data_state_seed::UInt64 : seed used to generate seed used to split data.

ProRF.view_reg3d — Function

view_reg3d(RI::AbstractRFI, Z::Matrix{Float64};
           title::Union{String, Nothing}=nothing,
           elev::Union{Real, Nothing}=nothing,
           azim::Union{Real, Nothing}=nothing,
           scale::Int=2)

Examples

julia> view_reg3d(RI, Z, title="NRMSE value", scale=3);

julia> view_reg3d(RI, MZ, title="NRMSE value", azim=90, scale=3);

julia> view_reg3d(RI, SZ, title="NRMSE SD value", elev=120, scale=3);

Arguments

RI::AbstractRFI : for only RFI.

Z::Matrix{Float64} : nrmse matrix.

title::Union{String, Nothing} : title of the 3d graph.

elev::Union{Real, Nothing} : elevation viewing angle.

azim::Union{Real, Nothing} : azimuthal viewing angle.

scale::Int : decimal place to determine the limitation value of z axis.