Toolbox

Documentaion for ProRF's useful things to perform data preprocessing or random forest.

Random Forest

ProRF.get_rf_value — Function

get_rf_value(X::Matrix{Float64}, Y::Vector{Float64};
             val_mode::Bool=false,
             iter::Int=10, test_size::Float64=0.3,
             feat_range::Int=4, base_tree::Int=50,
             memory_usage::Real=4.0,
             max_tree::Int=1000)

Examples

julia> NFeat, NTree, NDepth = get_rf_value(X, Y, iter=5, memory_usage=10, val_mode=true);

Find best three arguments for random forest.

Arguments

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

val_mode::Bool : when val_mode is true, function don't display anything.

iter::Int : number of operations iterations.

test_size::Float64 : size of test set.

feat_range::Int : scope of search for number of selected features.

base_tree::Int : number of trees used when navigating.

memory_usage::Real : available memory capacity (GB)

max_tree::Int : thresholds fornumber of trees for performance.

Return

opt_feat::Int : optimized number of selected features.

opt_tree::Int : optimized number of trees.

opt_depth::Int : optimized maximum depth of the tree.

ProRF.train_test_split — Function

train_test_split(X::Matrix{Float64}, Y::Vector{Float64};
                 test_size::Float64=0.3, 
                 data_state::UInt64=@seed)

Examples

julia> x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2);

Split X, Y data to train, test set.

Arguments

X::Matrix{Float64} : X data.

Y::Vector{Float64} : Y data.

test_size::Float64 : size of test set.

data_state::UInt64 : seed used to split data.

ProRF.nrmse — Function

nrmse(pre::Vector{Float64}, tru::Vector{Float64})

Compute normalized root mean square error with predict value and true value.

ProRF.test_nrmse — Function

test_nrmse(regr::RandomForestRegressor, X::Matrix{Float64}, Y::Vector{Float64})

Compute normalized root mean square error with regression model, X and Y data.

test_nrmse(regr::RandomForestRegressor, X::Matrix{Float64}, Y::Vector{Float64},
           data_state::UInt64;
           test_size::Float64=0.3, test_mode::Bool=true)

Compute test or train set normalized root mean square error with regression model, X, Y data and seed.

ProRF.min_max_norm — Function

min_max_norm(data::Vector{Float64})

Min-max normalization function.

ProRF.save_model — Function

save_model(model_loc::String, regr::RandomForestRegressor)

Examples

julia> M = rf_model(X, Y, 5, 300, val_mode=true);

julia> save_model("model.jld", M);
ERROR: model.jld is not .jld2 file

julia> save_model("model.jld2", M);

Save RandomForestRegressor model using JLD2, make sure filename extension set to .jld2.

ProRF.load_model — Function

load_model(model_loc::String)

Examples

julia> M = load_model("model.jld2");

julia> X, Y, L = get_data(R, 9, 'E');

julia> @printf "Total NRMSE : %.6f\n" test_nrmse(M, X, Y)
Total NRMSE : 0.136494

Load RandomForestRegressor model using JLD2, make sure filename extension set to .jld2.

ProRF.parallel_predict — Function

parallel_predict(regr::RandomForestRegressor, X::Matrix{Float64})

Execute DecisionTree.predict(regr, X) in parallel.

parallel_predict(regr::RandomForestRegressor, L::Vector{Int},
                 seq_vector::Vector{String};
                 convert::Dict{Char, Float64}=ProRF.volume)

Get raw sequence vector and L data to make X data and execute DecisionTree.predict(regr, X) in parallel.

Data preprocessing

ProRF.view_sequence — Function

view_sequence(fasta_loc::String, amino_loc::Int=1;
              fontsize::Int=9,
              seq_width::Int=800,
              save::Bool=false)

Display .fasta file sequence using PyCall.

Made it by referring to bokeh sequence aligner visualization program.

Arguments

fasta_loc::String : location of .fasta file.

amino_loc::Int : start index for .fasta file (when value is not determined, set to 1).

fontsize::Int : font size of sequence.

seq_width::Int : gap width between sequences.

save::Bool : save .html viewer.

view_sequence(R::AbstractRF;
              fontsize::Int=9,
              seq_width::Int=800,
              save::Bool=false)

Display .fasta file sequence exists at a location in the AbstractRF object using PyCall.

Made it by referring to bokeh sequence aligner visualization program.

Arguments

R::AbstractRF : for both RF and RFI.

fontsize::Int : font size of sequence.

seq_width::Int : gap width between sequences.

save::Bool : save .html viewer.

ProRF.view_mutation — Function

view_mutation(fasta_loc::String)

Analyze data from .fasta file location of AbstractRF, then draw histogram, line graph about the mutation distribution.

view_mutation(R::AbstractRF)

Analyze data from .fasta file, then draw histogram, line graph about the mutation distribution.

Convert dictionary

Convert dictionary for get_data

Amino acid	Molar volume [4]	pI [2]	Hydrophobicity [3]	Frequency [1]

Alanine	88.6	6.00	51	0.0777

Arginine	173.4	10.76	-144	0.0627

Asparagine	114.1	5.41	-84	0.0336

Aspartic acid	111.1	2.77	-78	0.0542

Cysteine	108.5	5.07	137	0.0078

Glutamic acid	138.4	3.22	-115	0.0859

Glutamine	143.8	5.65	-128	0.0315

Glycine	60.1	5.97	-13	0.0730

Histidine	153.2	7.59	-55	0.0192

Isoleucine	166.7	6.02	106	0.0666

Leucine	166.7	5.98	103	0.0891

Lysine	168.6	9.74	-205	0.0776

Methionine	162.9	5.74	73	0.0241

Phenylalanine	189.9	5.48	108	0.0361

Proline	112.7	6.30	-79	0.0435

Serine	89.0	5.68	-26	0.0466

Threonine	116.1	5.60	-3	0.0487

Tryptophan	227.8	5.89	69	0.0102

Tyrosine	193.6	5.66	11	0.0300

Valine	140.0	5.96	108	0.0817

Frequency of amino acid is used to estimate the value of unknown amino acid.

ProRF.volume — Constant

Convert dictionary about molar volume of amino acid.

ProRF.pI — Constant

Convert dictionary about pI of amino acid.

ProRF.hydrophobicity — Constant

Convert dictionary about hydrophobicity of amino acid.

Others

ProRF._julia_interactive — Constant

Check julia currently running is interactive or not.

You can change the value in julia_isinteractive.

ProRF.julia_isinteractive — Function

julia_isinteractive()

Check _julia_interactive value.

julia_isinteractive(x::Bool)

Set _julia_interactive value.

ProRF.@seed — Macro

Return UInt64 range integer MersenneTwister RNG object seed. keep in mind that when macro executed, the seed is initialized.

ProRF.@show_pyplot — Macro

When _julia_interactive is on, execute display(gcf()) or _julia_interactive is off, execute show() and wait until the user inputs enter.

Reference

[1] : Brooks, D. J., Fresco, J. R., Lesk, A. M., & Singh, M. (2002). Evolution of amino acid frequencies in proteins over deep time: inferred order of introduction of amino acids into the genetic code. Molecular Biology and Evolution, 19(10), 1645–1655. https://doi.org/10.1093/oxfordjournals.molbev.a003988

[2] : Lide, D. R. (Ed.). (1991). Hdbk of chemistry & physics 72nd edition (72nd ed.). CRC Press.

[3] : Naderi-Manesh, H., Sadeghi, M., Arab, S., & Moosavi Movahedi, A. A. (2001). Prediction of protein surface accessibility with information theory. Proteins, 42(4), 452–459. https://doi.org/10.1002/1097-0134(20010301)42:4<452::aid-prot40>3.0.co;2-q

[4] : Zamyatnin, A. (1984). Amino Acid, Peptide, and Protein Volume in Solution. Annual Review of Biophysics and Biomolecular Structure, 13(1), 145–165. https://doi.org/10.1146/annurev.biophys.13.1.145